+++ website/mdx/components/P.tsx
@@ -1,9 +1,5 @@
 import { Text, TextProps } from "@chakra-ui/layout";
 
 export const P = (props: TextProps) => {
-  if (typeof props.children !== "string") {
-    return <>{props.children}</>;
-  }
-
-  return <Text mb={8} fontSize="lg" lineHeight="taller" {...props} />;
+  return <Text fontSize="lg" lineHeight="taller" {...props} />;
 };